refactor: BaseDialogFragment
