+++ app/src/main/java/com/sdhong/jonbeowin/base/BaseDialogFragment.kt
@@ -1,0 +1,30 @@
+package com.sdhong.jonbeowin.base
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.DialogFragment
+import androidx.viewbinding.ViewBinding
+
+abstract class BaseDialogFragment<VB : ViewBinding>(
+    private val bindingFactory: (LayoutInflater, ViewGroup?, Boolean) -> VB
+) : DialogFragment() {
+
+    private var _binding: VB? = null
+    protected val binding get() = _binding!!
+
+    final override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = bindingFactory(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
