+++ app/src/main/java/com/sdhong/jonbeowin/feature/encourage/EncourageDialogFragment.kt
@@ -1,12 +1,11 @@
 package com.sdhong.jonbeowin.feature.encourage
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.viewModels
 import com.sdhong.jonbeowin.R
+import com.sdhong.jonbeowin.base.BaseDialogFragment
 import com.sdhong.jonbeowin.databinding.FragmentEncourageDialogBinding
 import com.sdhong.jonbeowin.feature.encourage.uistate.EncourageDialogUiState
 import com.sdhong.jonbeowin.feature.encourage.viewmodel.EncourageDialogViewModel
@@ -15,21 +14,11 @@ import com.sdhong.jonbeowin.util.collectLatestFlow
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class EncourageDialogFragment : DialogFragment() {
-
-    private var _binding: FragmentEncourageDialogBinding? = null
-    private val binding get() = _binding!!
+class EncourageDialogFragment : BaseDialogFragment<FragmentEncourageDialogBinding>(
+    bindingFactory = FragmentEncourageDialogBinding::inflate
+) {
 
     private val viewModel: EncourageDialogViewModel by viewModels()
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentEncourageDialogBinding.inflate(inflater, container, false)
-        return binding.root
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -113,8 +102,4 @@ class EncourageDialogFragment : DialogFragment() {
         }
     }
 
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
 }
